Refactor HFSS and DUMMY solve into separate file
